Extracted out the interfaces that were declared in the file UserManagementProcessor.ts file.
