+++ backend/src/types/IBetaUserForConversion.ts
@@ -1,0 +1,9 @@
+export interface IBetaUserForConversion {
+    id: number;
+    first_name: string;
+    last_name: string;
+    email: string;
+    company_name?: string;
+    phone_number?: string;
+    country?: string;
+}
